+++ DapperTrabalhoFinal/DapperTrabalhoFinal/Controllers/UsersController.cs
@@ -2,7 +2,6 @@
 using DapperTrabalhoFinal.Config;
 using DapperTrabalhoFinal.Models;
 using Microsoft.AspNetCore.Mvc;
-using Oracle.ManagedDataAccess.Client;
 
 namespace DapperTrabalhoFinal.Controllers
 {
@@ -12,30 +11,6 @@ namespace DapperTrabalhoFinal.Controllers
 
     public class UsersController
     {
-
-        [HttpGet("Conexao")]
-
-        public string TestarConexao()
-        {
-            Conexao c = new Conexao();
-
-            OracleConnection obj = c.RealizarConexao();
-
-            obj.Open();
-
-            string Mensagem;
-
-            if (obj.State.ToString() == "Open")
-            {
-                Mensagem = "Conexão efetuada";
-            }
-            else
-            {
-                Mensagem = "Falha ao cadastrar";
-            }
-
-            return Mensagem;
-        }
 
         [HttpGet]
 
@@ -56,7 +31,7 @@ namespace DapperTrabalhoFinal.Controllers
 
             using var connection = c.RealizarConexao();
 
-            connection.Execute(@"INSERT INTO users (user_name, user_email, user_password, user_description, user_link, user_socialmedia, user_experience) VALUES (:User_name, :User_email, :User_password, :User_description, :User_link, :User_socialmedia, :User_experience)", u);
+            connection.Execute(@"INSERT INTO users (user_name, user_email, user_password, user_description, user_link, user_socialmedia, user_profession, user_hours_week, user_experience) VALUES (:User_name, :User_email, :User_password, :User_description, :User_link, :User_socialmedia, :User_profession, :User_hours_week, :User_experience)", u);
 
             return "Cadastro efetuado com sucesso!";
         }
@@ -69,7 +44,7 @@ namespace DapperTrabalhoFinal.Controllers
 
             using var conncetion = c.RealizarConexao();
 
-            conncetion.Execute(@"UPDATE users SET user_name = :User_name, user_email = :User_email, user_password = :User_password, user_description = :User_description, user_link = :User_link, user_socialmedia = :User_socialmedia, user_experience = :User_experience WHERE id_user = :Id_user", u);
+            conncetion.Execute(@"UPDATE users SET user_name = :User_name, user_email = :User_email, user_password = :User_password, user_description = :User_description, user_link = :User_link, user_socialmedia = :User_socialmedia, user_profession = :User_profession, user_hours_week = :User_hours_week, user_experience = :User_experience WHERE id_user = :Id_user", u);
 
             return "Pessoa alterada com sucesso!";
         }
@@ -82,9 +57,26 @@ namespace DapperTrabalhoFinal.Controllers
 
             using var connection = c.RealizarConexao();
 
-            connection.Execute(@"DELETE FROM users WHERE id_user = " + id_user);
+            int count = contabilizar(id_user);
 
-            return "Pessoa removida com sucesso!";
+            if (count > 0)
+            {
+                connection.Execute(@"DELETE FROM users WHERE id_user = " + id_user);
+                return "Removido com sucesso!";
+            }
+            else
+            {
+                return $"Falha na remoção {count}";
+            }
+        }
+
+        private int contabilizar(int id)
+        {
+            Conexao c = new Conexao();
+
+            using var connection = c.RealizarConexao();
+
+            return connection.ExecuteScalar<int>(@"SELECT COUNT(*) FROM users WHERE id_user = " + id);
         }
 
     }
